adjust to new structure
