+++ auxiliary/producers/traindataproducer/BUILD
@@ -11,7 +11,7 @@ go_library(
     srcs = [
         "server.go"
     ],
-    importpath = "github.com/bptlab/cepta/aux/producers/traindataproducer",
+    importpath = "github.com/bptlab/cepta/auxiliary/producers/traindataproducer",
     visibility = ["//visibility:private"],
     deps = [
         ":livetraindatareplayer",
@@ -33,7 +33,7 @@ go_library(
     srcs = [
         "LiveTrainDataReplayer.go"
     ],
-    importpath = "github.com/bptlab/cepta/aux/producers/traindataproducer/livetraindatareplayer",
+    importpath = "github.com/bptlab/cepta/auxiliary/producers/traindataproducer/livetraindatareplayer",
     visibility = ["//visibility:private"],
     deps = [
         "//osiris/lib:db",
